Renamed directory from structure to tests.
