+++ app/src/androidTest/java/tests/grid/GridUnit.java
@@ -1,4 +1,4 @@
-package structure.grid;
+package tests.grid;
 
 import noteworthyengine.GridNode;
 import noteworthyframework.Unit;
