+++ app/src/androidTest/java/tests/reflection/ReflectionTest.java
@@ -1,4 +1,4 @@
-package structure.reflection;
+package tests.reflection;
 
 import android.app.Application;
 import android.test.ApplicationTestCase;
